isl_multi_templ.c: extract out isl_multi_splice_templ.c
This makes the function isl_multi_*_splice opt-in rather than opt-out.
